% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/act_fcts.R
\name{act_fcts}
\alias{act_fcts}
\title{List Current Activation Functions}
\usage{
act_fcts()
}
\value{
A data.frame containing the following columns:
\item{act_fct}{Code for each activation function}
\item{description}{Brief detail for each model}
}
\description{
The function provides a list of all currently implemented
activation functions and a brief description.
}

