% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{activation_scores_table}
\alias{activation_scores_table}
\title{Activation Scores Matrix}
\usage{
activation_scores_table(object, raw_x, ...)
}
\arguments{
\item{object}{An "nn_model" object returned from \code{\link{compute}}}

\item{raw_x}{The raw data initially submitted to \code{\link{neuralnet}}}

\item{d_var}{A character vector of the dependent variable column
name(s).}

\item{step}{A fraction denoting the steps for the quantiles
(e.g. 0.01 = 100 steps)}

\item{breaks}{Optional parameter to alter default quantile breaks}

\item{allowParallel}{Optional parameter to allow parallel processing}
}
\value{
A matrix containing the activation levels from the "nn_model" and
the calculated scores (i.e. p-values).
}
\description{
This function is designed to return a matrix
containing the activation levels and associated scores (i.e. p-values)
for a trained model.
}

