% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{assign_categories}
\alias{assign_categories}
\title{Assign Testing Data Categories}
\usage{
assign_categories(object, ...)
}
\arguments{
\item{object}{Object of class nn_model or matrix/big.matrix which contains the
necessary columns for assigning categories}
}
\value{
A character vector containing the assigned categories for the specific model
}
\description{
Takes the columns containing category identificatiers (i.e. 0 or 1) in the
test dataset and returns a vector of the assigned category for the specific model type.
For example, payor model returns categories 'C','M','D', & 'S'.
}

