% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/payor_model.R
\name{assign_categories_manual}
\alias{assign_categories_manual}
\title{Assign Testing Data Categories Manually}
\usage{
assign_categories_manual(x, model_type = NULL, dv = NULL)
}
\arguments{
\item{x}{A data.frame, matrix, or big.matrix that contains at required columns
that contain '0' or '1' to indicate presence of a category.  Designed to accept the
'net.result' object from the \code{\link{compute}} function.}
}
\value{
A character vector representing the four payor categories
}
\description{
Takes the columns containing payor category identification in a test dataset
and returns a vector of the assigned category (e.g. 'C','M','D','S').
}
\details{
It is assumed that the specific order of the columns corresponds to a specific
order with respect to the specific \code{model_type}.
\describe{
 \item{payor}{Commerical, Medicare, Medicaid, and Self-Pay}
}
}

