% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_model_functions.R
\name{assign_manual_quantile_scores}
\alias{assign_manual_quantile_scores}
\title{Assign Manual Quantile Scores}
\usage{
assign_manual_quantile_scores(x, ntiles, allowParallel = FALSE)
}
\arguments{
\item{x}{An n x 3 matrix containing the raw activation values from}

\item{ntiles}{The n-tiles for each confidence category activation values}

\item{allowParallel}{Boolean indication if parallelization should be used}
}
\value{
The 'x' matrix with the score columns filled
}
\description{
This function computes scores defined by provided cut points.
This functions requires the user to provide the breakpoints to manually
calculate the scores.  This is representative of the original approach to
calculating scores.
}

