% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binary_model.R
\name{assign_predicted_binary_category}
\alias{assign_predicted_binary_category}
\title{Assign Predicted Binary Category}
\usage{
assign_predicted_binary_category(x)
}
\arguments{
\item{x}{The result.net object from \code{\link{compute}}.}
}
\value{
A character vector representing the binary categories
}
\description{
This function takes the 'result.net' object from the
\code{\link{compute}} function
and returns a vector of the assigned binary category codes
(i.e. present/absent).
}

