% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{assign_predicted_categories}
\alias{assign_predicted_categories}
\title{Assign Predicted Categories}
\usage{
assign_predicted_categories(object)
}
\arguments{
\item{object}{A 'nn_model' or 'model_subset' object}
}
\value{
A character vector representing the predicted model categories
}
\description{
This function takes the 'nn_model' object from the
\code{\link{compute}} function.
and returns a vector of the specific category type according to the model type.
}
\seealso{
\code{\link{assign_predicted_payor_category}}
}

