% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_model_functions.R
\name{assign_quantile_category}
\alias{assign_quantile_category}
\title{Assign Quantile Categories}
\usage{
assign_quantile_category(x, cutoffs = NULL)
}
\arguments{
\item{x}{A matrix containing at least 3 columns corresponding to 'raw',
'score' and 'cat' columns.}

\item{cutoffs}{A numeric vector of length 8, specifying the cutoffs for
the scores to correspond to the respective category (1-7).}
}
\value{
The 'x' matrix with the category columns filled.
}
\description{
This function assigns confidence categories defined by the 
user specified n-tiles
}

