% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_model_functions.R
\name{assign_quantile_scores}
\alias{assign_quantile_scores}
\title{Assign Quantile Scores}
\usage{
assign_quantile_scores(x, ntiles, allowParallel = FALSE)
}
\arguments{
\item{x}{An n x 16 matrix containing the raw activation values from}

\item{ntiles}{The n-tiles for each payor category activation values}

\item{allowParallel}{Boolean indication if parallelization should be used}
}
\value{
The 'x' matrix with the category columns filled
}
\description{
This function computes scores defined by the confidence category,
user specified n-tiles
}

