% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{assign_scores_and_categories}
\alias{assign_scores_and_categories}
\title{Assign Scores and Confidence Categories}
\usage{
assign_scores_and_categories(object, raw_x, ...)
}
\arguments{
\item{object}{An 'nn_model' object}

\item{raw_x}{The raw data initially submitted to \code{\link{neuralnet}}}

\item{breaks}{Optional parameter to alter default quantile breaks}

\item{cutoffs}{Optional parameter to alter default category breaks}

\item{allowParallel}{Optional parameter to allow parallel processing}
}
\value{
A 'model_scores' object which contains:

Matrix containing the four original category columns, raw activation values,
calculated scores, and confidence categories
}
\description{
Wrapper to quickly report the confidence categories and
scaled scores for each sample from a neural net model.  This assigns scores
based on the raw quantile level.
}
\seealso{
\code{\link{assign_payor_scores_and_categories}}
}

