% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big.compute.R
\name{big.compute}
\alias{big.compute}
\title{"big.matrix" compatible computation of a given neural network for given covariate vectors}
\usage{
big.compute(x, covariate, rep = 1, model_type = NULL)
}
\arguments{
\item{x}{an object of class nn.}

\item{covariate}{a dataframe or matrix containing the variables that had been used to train the neural network.}

\item{rep}{an integer indicating the neural network's repetition which should be used.}

\item{model_type}{State specific model type that was fit (e.g. payor)}
}
\value{
\item{compute}{returns a list containing the following components:}

\item{neurons}{a list of the neurons' output for each layer of the neural network.}

\item{net.result}{a matrix containing the overall result of the neural network.}
}
\description{
compute, a method for objects of class nn, typically produced by neuralnet. Computes
the outputs of all neurons for specific arbitrary covariate vectors given a trained neural network.
Please make sure that the order of the covariates is the same in the new matrix or dataframe as in
the original neural network.
}
\details{
This is a replication of the \link[neuralnet]{compute} function only made compatible with big.matrix objects.
}

