% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binary_model.R
\name{binary_classification_curve}
\alias{binary_classification_curve}
\title{Create Classification Curve for Binary Model}
\usage{
binary_classification_curve(x, filename)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the classification curves
for a generic binary model corresponding to the probability of correctly
classified category with respect to quantile levels.
}

