% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_model.R
\name{binary_lift_curve}
\alias{binary_lift_curve}
\title{Create Lift Curve for Binary Model}
\usage{
binary_lift_curve(x, data, dvs, filename)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{data}{A data.frame or matrix containing the original dependent
variable column}

\item{dvs}{A string specifying the dependent variable}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the lift curves
for a generic binary model corresponding to the probability of correctly
classified category with respect to quantile levels.
}

