% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binary_model.R
\name{binary_score_subset}
\alias{binary_score_subset}
\title{Binary Model Score Subsetting}
\usage{
binary_score_subset(x, cutoff)
}
\arguments{
\item{x}{The output from \code{\link{assign_scores_and_categories}} of class
\code{binary_scores}.}

\item{cutoff}{The cutoff for filtering predictions (e.g. all those
with confidence category >= 5)}
}
\value{
\item{pred}{Subset of the neuralnet activations}

\item{raw}{Subset of the test data binary category}
}
\description{
Subset the output from scores_and_categories to evaluate
prediction metrics above a given cutoff (e.g. 4, 7, etc.).  This is for the
generic binary model with no customizations.
}

