% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{classification_curve}
\alias{classification_curve}
\title{Create Classification Curve for Classification Model}
\usage{
classification_curve(object, ...)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the classification curves
for models corresponding to the probability of correctly
classified categories with respect to quantile levels.
}

