% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{cor_pairs}
\alias{cor_pairs}
\title{Maximum correlations}
\usage{
cor_pairs(d)
}
\arguments{
\item{d}{A correlation matrix}
}
\description{
Returns which correlations in a correlation matrix are the highest.
}

