% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvControl.R
\name{cvControl}
\alias{cvControl}
\title{Control function for 'train'}
\usage{
cvControl(model_args, model_type, filter = FALSE, filter_category = NULL,
  verbose = FALSE, scale = FALSE)
}
\arguments{
\item{model_args}{A named list of additional arguements for the internal
method called.}

\item{filter}{Boolean argument indicating if prediction results should
be filtered}

\item{filter_category}{A numeric value for the category filter if \code{filter}
is \code{TRUE}}

\item{verbose}{A boolean option for verbose model output}

\item{scale}{A boolean option whether to scale model results
cross-validation step which can be subsequently loaded with \code{load}}
}
\value{
A list of the additional internal parameters
}
\description{
This function provides a means to consolidate additional
parameters for the train function to avoid verbose input parameters.
}

