% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linear_model_metrics.R
\name{delta_models}
\alias{delta_models}
\title{Delta Models}
\usage{
delta_models(base_model, new_iv)
}
\arguments{
\item{new_iv}{A string name of the new variable to be added}

\item{model}{A fitted lm model to be used as a base comparison}
}
\description{
Returns a dataframe of delta R2 and p-value of an added Independent Variable
to a base model.
}

