% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{demographic_profile}
\alias{demographic_profile}
\title{Demographic Profile}
\usage{
demographic_profile(object, ...)
}
\arguments{
\item{x}{A 'model_scores' object (i.e. the output from 
\code{\link{assign_scores_and_categories}})}

\item{data}{The raw data initially used to fit the model containing all variables.}
}
\value{
Matrix containing the sums of each column within each category of
each group.
}
\description{
This function is used for a naieve demographic profile
of each category within a scored and categorized model.  It returns
a matrix of size M x (7*num.groups).  The values within a simply the
sum of each column for each category within each group (as data is binary).
}
\note{
The 'data' parameter may also be a subset of columns which contains only
the variables of interest.  However, the subset must still contain the dependent
variables.  This is used as an internal check to make sure only data for the specific
model is evaluated.
}

