% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/denovo.grid.R
\name{denovo.grid}
\alias{denovo.grid}
\title{Denovo Grid Generation}
\usage{
denovo.grid(data, method, res)
}
\arguments{
\item{data}{data of method to be tuned}

\item{method}{vector indicating the models to generate grids.
Available options are \code{"neuralnet"} (Neural Network),
\code{"rf"} (Random Forest), \code{"gbm"} (Gradient Boosting Machine),
 \code{"svm"} (Support Vector Machines), and \code{"glmnet"}
 (Elastic-net Generalized Linear Model)}

\item{res}{Resolution of model optimization grid.}
}
\value{
A list containing dataframes of all combinations of
parameters for each model:
}
\description{
Greates grid for optimizing selected models
}
\author{
Charles Determan Jr
}
\seealso{
\code{"expand.grid"} for generating grids of specific
parameters desired.  However, NOTE that you must still convert
the generated grid to a list.
}

