% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/denovo.grid.R
\name{denovo_neuralnet_grid}
\alias{denovo_neuralnet_grid}
\title{Denovo Grid Generation}
\usage{
denovo_neuralnet_grid(res, act_fcts, dropout)
}
\arguments{
\item{res}{Resolution of the search grid}

\item{act_fcts}{A character vector of activation functions}

\item{dropout}{A boolean indicating if dropout iterations should be included}
}
\description{
Greates expanded grid for neuralnets
}
\author{
Charles Determan Jr
}
\seealso{
\code{\link{denovo.grid}} or \code{\link{"expand.grid"}} for generating grids
of specific parameters desired.  However, NOTE that you must still use the same
arguments names prefixed with a '.'.
}

