% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{extract_breakpoints}
\alias{extract_breakpoints}
\title{Extract Breakpoints}
\usage{
extract_breakpoints(object, ...)
}
\arguments{
\item{x}{A 'model_scores' object (i.e. the output from
\code{\link{assign_scores_and_categories}})}
}
\value{
A numeric vector containing the breakpoints
}
\description{
This function is used to extract the breakpoints
for a standard 7 category scored model.  It returns a vector
of values starting from 0 and ending with 1.
}
\note{
The vector returned does not include the absolute min/max from the
scored data.  This was intentional in order to be more inclusive
when the model is used to score.
}

