% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{gains_chart}
\alias{gains_chart}
\title{Create Gains Chart}
\usage{
gains_chart(object, ...)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the cumulative classification 
curves for models corresponding to quantile levels.
}

