% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{lift_curve}
\alias{lift_curve}
\title{Create Lift Curve for Classification Model}
\usage{
lift_curve(object, ...)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{data}{A data.frame or matrix containing the original dependent
variable column}

\item{dvs}{A string specifying the dependent variable}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the lift curves
for models corresponding to the total utilization with
respect to score quantile levels.
}

