% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linear_model_metrics.R
\name{mad_model}
\alias{mad_model}
\title{MAD - Mean Absolute Deviation}
\usage{
mad_model(model, raw_data, expo = TRUE)
}
\arguments{
\item{model}{A fitted model (e.g. lm, glm, etc.)}

\item{raw_data}{The results that you wish to compare the predicted results}

\item{expo}{Whether the results of the prediction should be raised to the exponent.
Such as when you fit a model with the log of the dependent variable. Default \code{"expo = TRUE"}.}
}
\description{
Returns the average difference between a model's prediction and
the actual testing data.
}

