% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.R
\name{metrics}
\alias{metrics}
\title{Performance Metrics}
\usage{
metrics()
}
\value{
A data.frame containing the metric codes
and descriptions
}
\description{
Function to list avaialble model performance
metrics to tune a model.
}

