% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_types.R
\name{model_types}
\alias{model_types}
\title{List Current Model Types}
\usage{
model_types()
}
\value{
A data.frame containing the following columns:
\item{model_types}{Code for each model type}
\item{description}{Brief detail for each model}
}
\description{
The function provides a list of all currently implemented models
and a brief description of each.
}

