% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelInfo.R
\name{modelInfo}
\alias{modelInfo}
\title{Model Info}
\usage{
modelInfo()
}
\value{
A data.frame containing:
\item{methods}{The abbreviated code for the method}
\item{description}{Full name of the method}
\item{hyperparams}{Comma separated string of hyperparameters for tuning}
}
\description{
Provide a list of available models and respective information
}
\author{
Dr. Charles Determan Jr. PhD
}

