% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiAUC.R
\name{multiAUC}
\alias{multiAUC}
\title{Multi-class AUC}
\usage{
multiAUC(classes, probs)
}
\arguments{
\item{classes}{A vector of 'observed' classes}

\item{probs}{A matrix containing a column for each class with respective probabilities}
}
\description{
This function is very similar to the normal \code{\link[pROC]{auc}} 
function in the \code{pROC} package.  It performs multiclass AUC as defined by 
Hand and Till (2001) whereby the AUC is calculated for each class and subsequently 
average.  This function extends this function to accept a probability matrix.
}
\references{
David J. Hand and Robert J. Till (2001). A Simple Generalisation of the 
Area Under the ROC Curve for Multiple Class Classification Problems. Machine Learning 
45(2), p. 171–186. DOI: 10.1023/A:1010920819831.
}

