% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payor_model.R
\name{payor_classification_curve}
\alias{payor_classification_curve}
\title{Create Classification Curve for Payor Model}
\usage{
payor_classification_curve(x, filename)
}
\arguments{
\item{x}{The output from \code{\link{percentile_profile}}}

\item{filename}{The name of the file to be saved as a pdf}
}
\description{
This function quickly creates the lift curves
for the payor model corresponding to the probability of correctly
classified payor category with respect to quantile levels.
}

