% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{percentile_profile}
\alias{percentile_profile}
\title{Probability Classification Profile}
\usage{
percentile_profile(object, ...)
}
\arguments{
\item{x}{The output from \code{\link{assign_scores_and_categories}}}

\item{step}{The width of the bins created from 0 to 1}

\item{allowParallel}{Optional argument to utilize parallel processing}
}
\description{
This function calculates the probability of the group
being accurately predicted across user defined bins of length \code{step}.
This provide a matrix of values that can be used to rapidly create a lift
curve with the \code{\link{lift_curve}} function.
}

