% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{predicting}
\alias{predicting}
\title{Model Group Prediction}
\usage{
predicting(modelFit, method, newdata, model_type = NULL, model_args = NULL,
  param = NULL, scale = FALSE)
}
\arguments{
\item{modelFit}{The fitted model being evaluated}

\item{method}{String of the model to be evaluated}

\item{newdata}{The testing data to predict group membership}

\item{model_args}{Additional arguments for individual models}

\item{param}{The parameters being fit to the model 
(Determined by model optimization).}
}
\value{
Returns a list of predicted group membership
}
\description{
This function evaluates a single fitted model and returns 
the predicted group memberships.
}

