% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.sql.R
\name{read.sql}
\alias{read.sql}
\title{Read SQL file}
\usage{
read.sql(dir)
}
\arguments{
\item{dir}{Location of the .sql file}
}
\description{
Reads a .sql file and creates a character vector to be used in a dbSendQuery().
}

