% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{scale01}
\alias{scale01}
\title{Scaling 0-1}
\usage{
scale01(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
a scaled numeric vector
}
\description{
This is just a small function to easily scale a range
of numbers to have a min = 0 and max = 1.
}

