% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{score_subset}
\alias{score_subset}
\title{Subset Score Categories}
\usage{
score_subset(object, cutoff, ...)
}
\arguments{
\item{object}{A 'model_scores' object}

\item{cutoff}{The cutoff for filtering predictions (e.g. all those with confidence category >= 5)}
}
\value{
A 'model_subset' object

\describe{
 \item{pred}{Subset of the neuralnet activations}
 \item{raw}{Subset of the test data category columns}
 }
}
\description{
This function subsets the output from \code{\link{assign_scores_and_categories}} 
to evaluate prediction metrics above a given cutoff (e.g. 4, 7, etc.)
}
\seealso{
\code{\link{score_subset,payor_scores-method}}
}

