% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{score_subset,payor_scores-method}
\alias{score_subset,payor_scores-method}
\title{Subset Payor Model Score Categories}
\usage{
\S4method{score_subset}{model_scores}(object, cutoff, payor_category = NULL)
}
\arguments{
\item{object}{The output from \code{\link{assign_scores_and_categories}}}

\item{cutoff}{The cutoff for filtering predictions (e.g. all those with confidence category >= 5)}

\item{category}{Optional argument to filter only on one payor category}
}
\value{
A 'payor_subset' object

\describe{
 \item{pred}{Subset of the neuralnet activations}
 \item{raw}{Subset of the test data payor category columns}
 }
}
\description{
Subset the output from scores_and_categories to evaluate
prediction metrics above a given cutoff (e.g. 4, 7, etc.)
}
\seealso{
\code{\link{score_subset}}
}

